fixed a bug with extraction limits
I identified a bug. Turns out that using traces from veloce manual with new approach could cause a situation that traces on the edge does not produce the most extreme extraction limits that I'm padding with zeros (i.e. x values below or above x dimension of the frame). Implemented a temporary fix, but in the future it would be better to trim traces so they don't extend out of frame.